src/test/java/LivroTest.java
import org.junit.jupiter.api.Test;

import java.time.Instant;
import java.time.LocalDateTime;
import java.time.ZoneOffset;

import static org.junit.jupiter.api.Assertions.*;

class LivroTest {

    @Test

    public void Test_Livro() {

        Instant data_nascimento_autor = LocalDateTime.parse("2018-10-15t20:30:00").toInstant(ZoneOffset.UTC);
        Autor atual = new Autor("Marcelo", "Honda", data_nascimento_autor, 1);

        Instant data_publicacao = LocalDateTime.parse("2018-10-15t20:30:00").toInstant(ZoneOffset.UTC);
        Livro livro_atual = new Livro
                ("Java", atual, data_publicacao, 1, 1, 1);

        assertEquals("Java", livro_atual.getTitulo_livro());
        assertEquals(atual, livro_atual.getAutor());
        assertEquals(data_publicacao, livro_atual.getData_publication());
        assertEquals(1, livro_atual.getEdicao());
        assertEquals(1, livro_atual.getId_biblioteca());
        assertEquals(1, livro_atual.getId_livro());

    }

}